+++ pulsar-functions/instance/src/main/java/org/apache/pulsar/functions/instance/InstanceUtils.java
@@ -20,9 +20,9 @@ package org.apache.pulsar.functions.instance;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static org.apache.commons.lang3.StringUtils.isEmpty;
-import static org.apache.pulsar.functions.utils.FunctionCommon.ComponentType.FUNCTION;
-import static org.apache.pulsar.functions.utils.FunctionCommon.ComponentType.SINK;
-import static org.apache.pulsar.functions.utils.FunctionCommon.ComponentType.SOURCE;
+import static org.apache.pulsar.functions.utils.ComponentType.FUNCTION;
+import static org.apache.pulsar.functions.utils.ComponentType.SINK;
+import static org.apache.pulsar.functions.utils.ComponentType.SOURCE;
 
 import lombok.experimental.UtilityClass;
 
@@ -31,6 +31,7 @@ import org.apache.pulsar.client.api.Schema;
 import org.apache.pulsar.functions.api.SerDe;
 import org.apache.pulsar.functions.proto.Function;
 import org.apache.pulsar.functions.sink.PulsarSink;
+import org.apache.pulsar.functions.utils.ComponentType;
 import org.apache.pulsar.functions.utils.Reflections;
 
 import net.jodah.typetools.TypeResolver;
@@ -88,7 +89,7 @@ public class InstanceUtils {
         }
     }
 
-    public FunctionCommon.ComponentType calculateSubjectType(Function.FunctionDetails functionDetails) {
+    public ComponentType calculateSubjectType(Function.FunctionDetails functionDetails) {
         Function.SourceSpec sourceSpec = functionDetails.getSource();
         Function.SinkSpec sinkSpec = functionDetails.getSink();
         if (sourceSpec.getInputSpecsCount() == 0) {
@@ -118,7 +119,7 @@ public class InstanceUtils {
                 functionDetails.getName());
     }
 
-    public static Map<String, String> getProperties(FunctionCommon.ComponentType componentType,
+    public static Map<String, String> getProperties(ComponentType componentType,
                                                     String fullyQualifiedName, int instanceId) {
         Map<String, String> properties = new HashMap<>();
         switch (componentType) {
